+++ prover/src/tables64/lt.rs
@@ -312,32 +312,6 @@ pub fn bus_interactions() -> Vec<BusInteraction> {
     ]
 }
 
-// =========================================================================
-// Helper functions for constraint computation
-// =========================================================================
-
-/// Convert lhs (DWordHHW) to DWordWL format.
-/// DWordHHW: [Word, Half, Half] -> DWordWL: [Word, Word]
-/// lo = lhs[0]
-/// hi = lhs[1] + 2^16 * lhs[2]
-#[inline]
-pub fn dwordhhw_to_dwordwl(word: u32, half1: u16, half2: u16) -> (u32, u32) {
-    let lo = word;
-    let hi = (half1 as u32) + ((half2 as u32) << 16);
-    (lo, hi)
-}
-
-/// Convert DWordHL to DWordWL format.
-/// DWordHL: [Half, Half, Half, Half] -> DWordWL: [Word, Word]
-/// lo = half[0] + 2^16 * half[1]
-/// hi = half[2] + 2^16 * half[3]
-#[inline]
-pub fn dwordhl_to_dwordwl(h0: u16, h1: u16, h2: u16, h3: u16) -> (u32, u32) {
-    let lo = (h0 as u32) + ((h1 as u32) << 16);
-    let hi = (h2 as u32) + ((h3 as u32) << 16);
-    (lo, hi)
-}
-
 /// Compute virtual carry[0] and carry[1] for the addition rhs + lhs_sub_rhs = lhs
 ///
 /// From spec:
@@ -604,96 +578,4 @@ pub fn lt_constraints(constraint_idx_start: usize) -> (Vec<LtConstraint>, usize)
         }),
     ];
     (constraints, idx)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_lt_unsigned_basic() {
-        let ops = vec![
-            LtOperation::new(5, 10, false),       // 5 < 10 unsigned -> true
-            LtOperation::new(10, 5, false),       // 10 < 5 unsigned -> false
-            LtOperation::new(5, 5, false),        // 5 < 5 unsigned -> false
-            LtOperation::new(0, 1, false),        // 0 < 1 unsigned -> true
-            LtOperation::new(u64::MAX, 0, false), // MAX < 0 unsigned -> false
-        ];
-
-        assert!(ops[0].compute_lt());
-        assert!(!ops[1].compute_lt());
-        assert!(!ops[2].compute_lt());
-        assert!(ops[3].compute_lt());
-        assert!(!ops[4].compute_lt());
-    }
-
-    #[test]
-    fn test_lt_signed_basic() {
-        let ops = vec![
-            LtOperation::new(5, 10, true),             // 5 < 10 signed -> true
-            LtOperation::new(10, 5, true),             // 10 < 5 signed -> false
-            LtOperation::new((-5i64) as u64, 5, true), // -5 < 5 signed -> true
-            LtOperation::new(5, (-5i64) as u64, true), // 5 < -5 signed -> false
-            LtOperation::new((-10i64) as u64, (-5i64) as u64, true), // -10 < -5 signed -> true
-        ];
-
-        assert!(ops[0].compute_lt());
-        assert!(!ops[1].compute_lt());
-        assert!(ops[2].compute_lt());
-        assert!(!ops[3].compute_lt());
-        assert!(ops[4].compute_lt());
-    }
-
-    #[test]
-    fn test_trace_generation() {
-        let ops = vec![
-            LtOperation::new(100, 200, false),
-            LtOperation::new(200, 100, true),
-        ];
-
-        let trace = generate_lt_trace(&ops);
-
-        // Should be padded to power of 2
-        assert_eq!(trace.main_table.height, 2);
-        assert_eq!(trace.main_table.width, cols::NUM_COLUMNS);
-
-        // Check first row
-        let row0 = trace.main_table.get_row(0);
-        assert_eq!(row0[cols::LHS_0], FE::from(100u64));
-        assert_eq!(row0[cols::RHS_0], FE::from(200u64));
-        assert_eq!(row0[cols::SIGNED], FE::zero());
-        assert_eq!(row0[cols::LT], FE::one()); // 100 < 200
-
-        // Check second row
-        let row1 = trace.main_table.get_row(1);
-        assert_eq!(row1[cols::LHS_0], FE::from(200u64));
-        assert_eq!(row1[cols::RHS_0], FE::from(100u64));
-        assert_eq!(row1[cols::SIGNED], FE::one());
-        assert_eq!(row1[cols::LT], FE::zero()); // 200 < 100 signed -> false
-    }
-
-    #[test]
-    fn test_dwordhhw_conversion() {
-        // Test: 0x123456789ABCDEF0
-        // DWordHHW: [0x9ABCDEF0, 0x5678, 0x1234]
-        let (lo, hi) = dwordhhw_to_dwordwl(0x9ABC_DEF0, 0x5678, 0x1234);
-        assert_eq!(lo, 0x9ABC_DEF0);
-        assert_eq!(hi, 0x1234_5678);
-    }
-
-    #[test]
-    fn test_dwordhl_conversion() {
-        // Test: 0x123456789ABCDEF0
-        // DWordHL: [0xDEF0, 0x9ABC, 0x5678, 0x1234]
-        let (lo, hi) = dwordhl_to_dwordwl(0xDEF0, 0x9ABC, 0x5678, 0x1234);
-        assert_eq!(lo, 0x9ABC_DEF0);
-        assert_eq!(hi, 0x1234_5678);
-    }
-
-    #[test]
-    fn test_bus_interactions_count() {
-        let interactions = bus_interactions();
-        // MSB16 x2 + IS_HALFWORD x4 + LT x1 = 7 interactions
-        assert_eq!(interactions.len(), 7);
-    }
 }
