+++ prover/src/tests/mod.rs
@@ -8,3 +8,5 @@ pub mod bitwise_bus_tests;
 pub mod bitwise_tests;
 #[cfg(test)]
 pub mod constraints64_tests;
+#[cfg(test)]
+pub mod lt_tests;
